requirements.txt
Django==3.0.9
djangorestframework==3.11.0
psycopg2-binary==2.8.5
django-extra-fields==2.0.5
Pillow==7.2.0
